\name{SLRpg-package}
\alias{SLRpg-package}
\alias{SLRpg}
\docType{package}
\title{
\packageTitle{SLRpg}
}
\description{
\packageDescription{SLRpg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SLRpg}
\packageIndices{SLRpg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SLRpg}

Maintainer: \packageMaintainer{SLRpg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
